requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.0
django-embed-video==1.4.0
django-grappelli==2.15.1
django-taggit==1.5.1
fake-headers==1.0.2
fake-useragent==0.1.11
future==0.18.2
html5lib==1.1
idna==3.2
kinopoiskpy==0.8.2
kombu==5.1.0
lxml==4.6.3
Pillow==8.3.2
pkg_resources==0.0.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
Unidecode==1.3.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
